web_app/service/requirements.txt
boto3==1.11.16
Flask==1.1.1
Flask-Cors==3.0.8
werkzeug<1.0
mysql-connector-python==8.0.23